Refactor Steam identity to manual service with local storage
Replace server-side SteamIdentityService with ManualSteamIdentityService that uses local storage for Steam ID and vanity URL. Move LocalStorageKeys to shared contracts namespace for broader access. Update Home.razor to restore state without calling identity service on initialization. Add HttpClient configuration for API calls.
